+++ src/main/java/com/example/ControleAlmoxarifado/dto/material/CriacaoMaterialRespostaDTO.java
@@ -1,9 +1,11 @@
 package com.example.ControleAlmoxarifado.dto.material;
 
+import java.math.BigDecimal;
+
 public record CriacaoMaterialRespostaDTO(
         Long id,
         String nome,
         String unidade,
-        double estoque
+        BigDecimal estoque
 ) {
 }
